api/requirements.txt
aioredis==2.0.1
aniso8601==9.0.1
asgi-redis==1.4.3
asgiref==3.7.2
asn1crypto==1.5.1
async-timeout==4.0.3
attrs==23.1.0
autobahn==23.6.2
Automat==22.10.0
cffi==1.16.0
channels==4.0.0
channels-redis==4.1.0
constantly==15.1.0
cryptography==41.0.5
Cython==3.0.4
daphne==4.0.0
Django==4.2.6
django-cors-headers==4.3.0
django-graphql-jwt==0.4.0
Faker==19.12.0
graphene==3.3
graphene-django==3.1.5
graphql-core==3.2.3
graphql-relay==3.2.0
hiredis==2.2.3
hyperlink==21.0.0
idna==3.4
incremental==22.10.0
msgpack==1.0.7
msgpack-python==0.5.6
promise==2.3
psycopg2-binary==2.9.9
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
PyHamcrest==2.1.0
PyJWT==2.8.0
pyOpenSSL==23.3.0
python-dateutil==2.8.2
pytz==2023.3.post1
redis==5.0.1
Rx==3.2.0
service-identity==23.1.0
singledispatch==4.1.0
six==1.16.0
sqlparse==0.4.4
text-unidecode==1.3
Twisted==23.8.0
txaio==23.1.1
typing_extensions==4.8.0
zope.interface==6.1
